# docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Athlete:
    properties:
      age:
        description: The age of the athlete
        format: uint8
        minimum: 18
        type: integer
        x-go-name: Age
      athlete_id:
        description: The unique ID that distinguishes athletes
        readOnly: true
        type: string
        x-go-name: Athlete_ID
      first_name:
        description: The first name of the athlete
        type: string
        x-go-name: First_Name
      joined:
        description: The date the athlete joined AER
        readOnly: true
        type: string
        x-go-name: Joined
      last_name:
        description: The last name of the athlete
        type: string
        x-go-name: Last_Name
    required:
    - first_name
    - last_name
    - age
    type: object
    x-go-package: github.com/BrandonReno/A.E.R/docs
  Efficiency:
    properties:
      efficiency_score:
        description: The efficiency score of an athlete based upon their past workouts
        example: 20.34
        format: double
        readOnly: true
        type: number
        x-go-name: Efficiency_Score
      favorite_excercise:
        description: The favorite sport of the athlete based upon what they do in
          workouts most often
        example: swimming
        readOnly: true
        type: string
        x-go-name: Favorite_Sport
    type: object
    x-go-package: github.com/BrandonReno/A.E.R/docs
  Workout:
    properties:
      date:
        description: The date the workout was completed
        type: string
        x-go-name: Date
      description:
        description: The description of the workout, how the athlete felt, what they
          did, etc
        type: string
        x-go-name: Description
      rating:
        description: The athletes rating of the workout
        format: int64
        maximum: 10
        minimum: 0
        type: integer
        x-go-name: Rating
      sport:
        description: The sport/excercise the athlete did
        type: string
        x-go-name: Sport
      workout_id:
        description: The unique integer ID of an athletes workout
        format: int64
        readOnly: true
        type: integer
        x-go-name: Workout_ID
    required:
    - date
    - description
    - sport
    - rating
    type: object
    x-go-package: github.com/BrandonReno/A.E.R/docs
info:
  description: |-
    The purpose of this API is to handle back end
    AER services following RESTful principles
  title: 'AER: Workout API'
  version: 1.0.0
paths:
  /:
    get:
      description: Lists workouts from all athletes accross the service
      operationId: getAllWorkouts
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/workoutsResponse'
      schemes:
      - http
      tags:
      - workouts
  /athletes:
    get:
      description: Lists all athletes registered in the system
      operationId: listAthletes
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/athletesResponse'
      schemes:
      - http
      tags:
      - athletes
    post:
      consumes:
      - application/json
      description: Adds a new workout to the database
      operationId: addAthlete
      parameters:
      - in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Athlete'
      responses:
        "201":
          $ref: '#/responses/noContent'
        "400":
          $ref: '#/responses/verror'
      schemes:
      - http
      tags:
      - athletes
  /athletes/{athlete_id}:
    delete:
      description: Deletes a specified athlete by id
      operationId: deleteAthlete
      parameters:
      - example: H3bfj78eHe
        in: path
        name: athlete_id
        required: true
        type: string
        x-go-name: Athlete_id
      responses:
        "201":
          $ref: '#/responses/noContent'
        "404":
          $ref: '#/responses/badRequest'
      schemes:
      - http
      tags:
      - athletes
    get:
      description: Lists information from a specified athlete
      operationId: getAthlete
      parameters:
      - example: H3bfj78eHe
        in: path
        name: athlete_id
        required: true
        type: string
        x-go-name: Athlete_id
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/singleAthlete'
        "404":
          $ref: '#/responses/badRequest'
      schemes:
      - http
      tags:
      - athletes
    post:
      consumes:
      - application/json
      description: Adds a new workout to the database
      operationId: addWorkout
      parameters:
      - example: H3bfj78eHe
        in: path
        name: athlete_id
        required: true
        type: string
        x-go-name: Athlete_id
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/Workout'
      responses:
        "201":
          $ref: '#/responses/noContent'
        "400":
          $ref: '#/responses/verror'
      schemes:
      - http
      tags:
      - workouts
    put:
      consumes:
      - application/json
      description: Updates an athlete in the system based on the given ID
      operationId: updateAthlete
      parameters:
      - example: H3bfj78eHe
        in: path
        name: athlete_id
        required: true
        type: string
        x-go-name: Athlete_id
      - in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Athlete'
      produces:
      - application/json
      responses:
        "201":
          $ref: '#/responses/noContent'
        "400":
          $ref: '#/responses/verror'
        "404":
          $ref: '#/responses/badRequest'
      schemes:
      - http
      tags:
      - athletes
  /athletes/{athlete_id}/{workout_id}:
    delete:
      description: Deletes a specified workout by id
      operationId: deleteWorkout
      parameters:
      - example: 374638
        format: int64
        in: path
        name: workout_id
        required: true
        type: integer
        x-go-name: Workout_id
      - example: H3bfj78eHe
        in: path
        name: athlete_id
        required: true
        type: string
        x-go-name: Athlete_id
      responses:
        "201":
          $ref: '#/responses/noContent'
        "404":
          $ref: '#/responses/badRequest'
      schemes:
      - http
      tags:
      - workouts
  /athletes/{athlete_id}/aer:
    get:
      description: Shows the athletes current AER statistics
      operationId: getEfficiency
      parameters:
      - example: H3bfj78eHe
        in: path
        name: athlete_id
        required: true
        type: string
        x-go-name: Athlete_id
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/efficiencyResponse'
        "404":
          $ref: '#/responses/badRequest'
      schemes:
      - http
      tags:
      - efficiency
  /athletes/{athlete_id}/workouts:
    get:
      description: Lists workouts from a particular registered athlete
      operationId: getUserWorkouts
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/workoutsResponse'
        "404":
          $ref: '#/responses/badRequest'
      schemes:
      - http
      tags:
      - workouts
  /athletes/{athlete_id}/workouts/{workout_id}:
    get:
      description: Gets a single workout from a specified athlete
      operationId: getSingleWorkout
      parameters:
      - example: 374638
        format: int64
        in: path
        name: workout_id
        required: true
        type: integer
        x-go-name: Workout_id
      - example: H3bfj78eHe
        in: path
        name: athlete_id
        required: true
        type: string
        x-go-name: Athlete_id
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/singleWorkout'
        "404":
          $ref: '#/responses/badRequest'
      schemes:
      - http
      tags:
      - workouts
    put:
      consumes:
      - application/json
      description: Updates a workout in the system based on the given ID
      operationId: updateWorkout
      parameters:
      - example: 374638
        format: int64
        in: path
        name: workout_id
        required: true
        type: integer
        x-go-name: Workout_id
      - example: H3bfj78eHe
        in: path
        name: athlete_id
        required: true
        type: string
        x-go-name: Athlete_id
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/Workout'
      produces:
      - application/json
      responses:
        "201":
          $ref: '#/responses/noContent'
        "400":
          $ref: '#/responses/verror'
        "404":
          $ref: '#/responses/badRequest'
      schemes:
      - http
      tags:
      - workouts
produces:
- application/json
responses:
  athletesResponse:
    description: All athletes are returned in the response
    schema:
      items:
        $ref: '#/definitions/Athlete'
      type: array
  badRequest:
    description: Index Not Found
    headers:
      error:
        example: Could not find the workout or athlete ID in the database
        type: string
  efficiencyResponse:
    description: An athletes aer is returned in the response
    schema:
      $ref: '#/definitions/Efficiency'
  noContent:
    description: No content is returned
  singleAthlete:
    description: A single athlete returns in the response
    schema:
      $ref: '#/definitions/Athlete'
  singleWorkout:
    description: A single workout returns in the response
    schema:
      $ref: '#/definitions/Workout'
  verror:
    description: Validation Error
    headers:
      error:
        example: Unable to validate athlete or workout object
        type: string
  workoutsResponse:
    description: A list of workouts returns in the response
    schema:
      items:
        $ref: '#/definitions/Workout'
      type: array
schemes:
- http
swagger: "2.0"